+++ lib/headlines.rb
@@ -9,6 +9,7 @@ require "active_model_serializers"
 require "nokogiri"
 require_relative "headlines/domains_archive"
 require_relative "headlines/scan_domains/runner"
+require_relative "headlines/scan_domains/scanner"
 require_relative "headlines/security_headers/security_header"
 require_relative "headlines/security_headers/strict_transport_security"
 require_relative "headlines/security_headers/x_content_type_options"
