+++ plugin.rb
@@ -12,6 +12,7 @@ gem("faraday_middleware", "0.9.1")
 gem("rubyzip", "1.1.7", require_name: "zip")
 gem("upsert", "2.1.0")
 gem("ruby-progressbar", "1.7.5")
+gem("concurrent-ruby", "1.0.0", require_name: "concurrent")
 
 register_asset("stylesheets/base/variables.css.scss")
 register_asset("stylesheets/components/breadcrumb.css.scss")
